Authorization - Task Ownership

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { Task } from './entities/task.entity';
@@ -11,10 +12,12 @@ export class TasksService {
   constructor(
     @InjectRepository(TaskRepository)
     private taskRepository: TaskRepository,
-  ) {}
+  ) { }
 
-  async getTasks(filterDto: GetTasksFilterDto) {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(
+    filterDto: GetTasksFilterDto,
+    user: User) {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
   /*getTaskWithFilters(filterDto: GetTasksFilterDto): Task[] {
@@ -34,29 +37,45 @@ export class TasksService {
         return task;
     }*/
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(
+    id: number,
+    user: User,
+  ): Promise<Task> {
+
+    console.log(user.id);
+    //const found = await this.taskRepository.findOne(id);
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!found) {
       throw new NotFoundException(`Task with ID '${id}' not found`);
     }
+
     return found;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User
+  ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(
+    id: number,
+    user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     console.log(result);
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID '${id}' not found`);
     }
   }
 
-  async updateTaskStatusById(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatusById(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
